Commit #6: 12/18/25
* Worked on:
- Added response entities and responses for http requests for testing.
- Finally passing all Insomnia tests
- Partially fixed search function on website. When min and max are set to 0, $0 games do not show up. May need to mess with search query.

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -1,9 +1,11 @@
 package org.yearup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.data.mysql.MySqlCategoryDao;
@@ -12,11 +14,6 @@ import org.yearup.models.Product;
 
 import java.util.List;
 
-// add the annotations to make this a REST controller
-// add the annotation to make this controller the endpoint for the following url
-    // http://localhost:8080/categories
-// add annotation to allow cross site origin requests
-
 @RestController
 @RequestMapping("categories")
 @CrossOrigin
@@ -24,9 +21,6 @@ public class CategoriesController
 {
     private final CategoryDao categoryDao;
     private final ProductDao productDao;
-
-
-    // create an Autowired controller to inject the categoryDao and ProductDao
 
     @Autowired
     public CategoriesController(CategoryDao categoryDao, ProductDao productDao) {
@@ -40,52 +34,88 @@ public class CategoriesController
         return ResponseEntity.ok(categoryDao.getAllCategories());
     }
 
-    @GetMapping(path="/{id}")
-    public ResponseEntity<Category> getById(@PathVariable int id)
-    {
-        return ResponseEntity.ok(categoryDao.getById(id));
+    @GetMapping("/{id}")
+    public Category getById(@PathVariable int id) {
+
+        Category category = null;
+
+        try {
+            category = categoryDao.getById(id);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
+
+        if (category == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found.");
+        }
+        return category;
     }
 
-    // the url to return all products in category 1 would look like this
-    // https://localhost:8080/categories/1/products
-    @GetMapping(path="/{categoryId}/products")
+    @GetMapping("/{categoryId}/products")
     public ResponseEntity<List<Product>> getProductsById(@PathVariable int categoryId)
     {
         List<Product> products = productDao.listByCategoryId(categoryId);
-
+        if (products == null || products.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(products);
     }
-
-    // add annotation to call this method for a POST action
-    // add annotation to ensure that only an ADMIN can call this function
 
     @PostMapping
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Category> addCategory(@RequestBody Category category)
     {
-        return ResponseEntity.ok(categoryDao.create(category));
+        try
+        {
+            Category created = categoryDao.create(category);
+            return ResponseEntity.status(HttpStatus.CREATED).body(created);
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
-    // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
-    // add annotation to ensure that only an ADMIN can call this function
-
-    @PutMapping(path="/{id}")
+    @PutMapping("/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Void> updateCategory(@PathVariable int id, @RequestBody Category category)
     {
-        categoryDao.update(id, category);
-        return ResponseEntity.ok().build();
+        try
+        {
+            Category selected = categoryDao.getById(id);
+            if (selected == null) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
 
+            categoryDao.update(id, category);
+            return ResponseEntity.noContent().build();
+
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
-    // add annotation to call this method for a DELETE action - the url path must include the categoryId
-    // add annotation to ensure that only an ADMIN can call this function
-
-    @DeleteMapping(path="/{id}")
+    @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<Void> deleteCategory(@PathVariable int id)
     {
-        categoryDao.delete(id);
-        return ResponseEntity.noContent().build();
+        try
+        {
+            Category selected = categoryDao.getById(id);
+
+            if(selected == null) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+
+            categoryDao.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 }
